Started Vuetify form for editing profile
Created Vuetify form for editing profile at the bottom of the page. Date picker does not work - has to be type in manually as calendar does not pop up. Edits can not be submitted currently and the form is always visible - both things will be dealt with in next session.