Now uses Vuetify validation
Set up Vuetify validation for the form. Checks these through individual validation functions in the profile controller class. Is updated by the user model class but causes an issue in which the first, last and middle name attributes show up twice in user data due to different spellings of the key names - must be addressed.

#story[176]