Fixed build failure, changed error evaluation order
Removed an unreachable line of code which caused the build to fail.
Changed the order in which errors/forbidden actions are checked for so that the error responses are more logical (doesn't tell a user that they reached the limit if they are not trying to send any new emails).

#story[175]