Jest testing for form validation, bug fixes
Wrote tests with Jest for the form validator methods. Filled in any missing implementation and did refactoring of the validator methods when issues arose so that the tests pass (mostly).
Currently, two date of birth tests fail as dates such as 31/04 and 30/02 are taken as valid dates (which they are not).
Also, fixed some small bugs which I came across in the Homepage and editing functions as I was refactoring.

#pair[jsv22, oam23] #story[176]