Authentication and API spec update for NewEmail
In NewEmailController, updated so that authentication is now done before emails are updated for POST endpoint. Includes allowing admins to update another user's email.
Updated POST endpoint to fit the API spec on Learn as it sent request body in different format. (Fixes on both front end and back end.)

#story[242]