PUT /emails request fits REST API spec, fixed bug with registered emails
Altered front and back end PUT /profiles/{profileId}/emails implementation so that what is sent matches the REST API specification.
Updated some of the return statements so that they instead throw exceptions to be handled by the exception handlers rather than returning ResponseEntity objects themselves. For this, also wrote custom exceptions for when email is already registered to another user and for when the user tries to register too many emails.
Fixed bug where, if an email in the list was already registered to another user, it overwrote this and registered the email to the current user instead.

#story[226]