Created properties.service file, admin mode on/off on profile page
Moved the admin mode checking (getting/setting the cookies) into a properties.service file so the user interface is not dependent on the details of the implementation (using cookies).
Also, implemented the admin mode on/off in viewing profile pages so that admins are able to act as regular users while in admin mode.

#story[295]