Added delete account button on profile edit page
Added a delete profile button on the edit profile page. Causes modal to pop up which prompts user to confirm deletion.
Also, fixed an issue with the admin mode/permission levels which caused name and log out to not be shown when viewing a profile as an admin not in admin mode.
Made working with cookies more reusable.

#story[295]