Created PUT /profiles/profileId/role endpoint with tests
Created the PUT /profiles/{profileId}/role endpoint which allows admins to promote and demote users. Currently, uses a HashMap of admin role names to corresponding permission levels to store the role names. This might be best stored in the database (not in a helper method in the code).
Also, wrote some tests for the endpoint using Mockito.

#story[295]