Implemented HomeFeedController, HomeFeedResponse
Implemented HomeFeedController with function for GET /homefeed to get updates to activities the user had subscribed to.
Created a HomeFeedResponse class to use to format the home feed JSON sent in the HTTP response so that it fits our API spec.
Also, implemented query to use in retrieving the activity changes from the change log.

#story[376]