Remove result button, only allow adding result if not already added
Added a button to allow users to remove participant results. Currently not sending requests as this is a separate Agilefant task.
Also, now only shows outcomes with no participant result logged against it in the "Add new result" section of the UI so the users can only add one result per outcome.

#story[390]