Reworked the implementation of the saving of pins
Misha and i reworked the validation for the activitypin saving location. this was important because we were intiailly considering to leave the validation of all coordinates to the frontend, whihc is bad practise. currently untested but works so far. will do tests in the near future #implement #pair[jmy39, mmo155] #story[416]